# contracts/access/v1/access.openapi.json
{
  "components": {
    "schemas": {
      "AccessDecision": {
        "additionalProperties": false,
        "properties": {
          "allowed": {
            "const": true,
            "default": true,
            "title": "Allowed",
            "type": "boolean"
          },
          "capability": {
            "const": "forecast:read",
            "default": "forecast:read",
            "title": "Capability",
            "type": "string"
          },
          "principal_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
            "title": "Principal Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/ForecastCheckRequest"
          }
        },
        "required": [
          "principal_id",
          "scope"
        ],
        "title": "AccessDecision",
        "type": "object"
      },
...
      "ForecastCheckRequest": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "enum": [
              "store",
              "online"
            ],
            "title": "Channel",
            "type": "string"
          },
          "product_ids": {
            "items": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Product Ids",
            "type": "array"
          },
          "schema_version": {
            "const": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "selling_location_ids": {
            "items": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
              "type": "string"
            },
            "maxItems": 5,
            "minItems": 1,
            "title": "Selling Location Ids",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "product_ids",
          "selling_location_ids",
          "channel"
        ],
        "title": "ForecastCheckRequest",
        "type": "object"
      },
...
      "IdentityResponse": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "enum": [
                "forecast:read",
                "access:admin",
                "knowledge:read",
                "knowledge:index",
                "model:decide",
                "forecast:run"
              ],
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "knowledge_scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeResourceScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "principal_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
            "title": "Principal Id",
            "type": "string"
          },
          "roles": {
            "items": {
              "enum": [
                "viewer",
                "operator",
                "admin",
                "promoter",
                "pipeline"
              ],
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceScope"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "principal_id",
          "roles",
          "capabilities",
          "scope"
        ],
        "title": "IdentityResponse",
        "type": "object"
      },
...
      "KnowledgeResourceScope": {
        "additionalProperties": false,
        "properties": {
          "access_classes": {
            "items": {
              "enum": [
                "public_project",
                "project_internal",
                "restricted"
              ],
              "type": "string"
            },
            "maxItems": 3,
            "minItems": 1,
            "title": "Access Classes",
            "type": "array"
          },
          "document_statuses": {
            "items": {
              "enum": [
                "specified",
                "implemented",
                "verified",
                "deprecated",
                "historical"
              ],
              "type": "string"
            },
            "maxItems": 5,
            "minItems": 1,
            "title": "Document Statuses",
            "type": "array"
          },
          "environment": {
            "enum": [
              "local",
              "test"
            ],
            "title": "Environment",
            "type": "string"
          },
          "repositories": {
            "items": {
              "enum": [
                "Oskar-Stachowski/retailops-cloud-native-platform",
                "Oskar-Stachowski/retailops-ai-intelligence"
              ],
              "type": "string"
            },
            "maxItems": 2,
            "minItems": 1,
            "title": "Repositories",
            "type": "array"
          }
        },
        "required": [
          "environment",
          "repositories",
          "access_classes",
          "document_statuses"
        ],
        "title": "KnowledgeResourceScope",
        "type": "object"
      },
...
      "PolicyMetadata": {
        "additionalProperties": false,
        "properties": {
          "credential_count": {
            "maximum": 64.0,
            "minimum": 1.0,
            "title": "Credential Count",
            "type": "integer"
          },
          "policy_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
            "title": "Policy Id",
            "type": "string"
          },
          "principal_count": {
            "maximum": 32.0,
            "minimum": 1.0,
            "title": "Principal Count",
            "type": "integer"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "policy_id",
          "principal_count",
          "credential_count"
        ],
        "title": "PolicyMetadata",
        "type": "object"
      },
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "anyOf": [
              {
                "enum": [
                  "configuration-not-approved",
                  "idempotency-conflict",
                  "index-run-not-found",
                  "index-not-configured",
                  "queue-full",
                  "worker-busy",
                  "claim-lost",
                  "index-report-not-found"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "database-not-ready",
                  "scope-denied",
                  "batch-scope-limit",
                  "forecast-run-not-found",
                  "forecast-run-denied",
                  "invalid-idempotency-key",
                  "idempotency-conflict",
                  "input-from-future",
                  "model-not-approved",
                  "model-decision-incomplete",
                  "input-not-prepared",
                  "input-origin-mismatch",
                  "queue-full",
                  "input-coverage-mismatch",
                  "run-not-active"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "correlation_id": {
            "format": "uuid",
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "pattern": "^urn:uuid:[0-9a-f-]{36}$",
            "title": "Instance",
            "type": "string"
          },
          "readiness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Ready"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "const": "about:blank",
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "instance",
          "correlation_id"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "Ready": {
        "additionalProperties": false,
        "properties": {
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencyStatus"
            },
            "title": "Dependencies",
            "type": "array"
          },
          "role": {
            "default": "foundation",
            "enum": [
              "foundation",
              "ai_api"
            ],
            "title": "Role",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "ready",
              "degraded",
              "not_ready"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "dependencies"
        ],
        "title": "Ready",
        "type": "object"
      },
...
      "ResourceScope": {
        "additionalProperties": false,
        "properties": {
          "channels": {
            "items": {
              "enum": [
                "store",
                "online"
              ],
              "type": "string"
            },
            "maxItems": 2,
            "minItems": 1,
            "title": "Channels",
            "type": "array"
          },
          "product_ids": {
            "items": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
              "type": "string"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Product Ids",
            "type": "array"
          },
          "selling_location_ids": {
            "items": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Selling Location Ids",
            "type": "array"
          }
        },
        "required": [
          "product_ids",
          "selling_location_ids",
          "channels"
        ],
        "title": "ResourceScope",
        "type": "object"
      },
...
      "RetrievalRequest": {
        "additionalProperties": false,
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/RetrievalFilters"
          },
          "max_context_tokens": {
            "default": 6000,
            "maximum": 6000.0,
            "minimum": 1.0,
            "title": "Max Context Tokens",
            "type": "integer"
          },
          "purpose": {
            "default": "documentation",
            "enum": [
              "documentation",
              "implementation",
              "verified_state",
              "history"
            ],
            "title": "Purpose",
            "type": "string"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "top_k": {
            "default": 5,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "question"
        ],
        "title": "RetrievalRequest",
        "type": "object"
      },
      "RetrievalResult": {
        "additionalProperties": false,
        "properties": {
          "answer_generation": {
            "const": "not_implemented",
            "default": "not_implemented",
            "title": "Answer Generation",
            "type": "string"
          },
          "content_trust": {
            "const": "untrusted_reference",
            "default": "untrusted_reference",
            "title": "Content Trust",
            "type": "string"
          },
          "context_bytes": {
            "maximum": 24000.0,
            "minimum": 0.0,
            "title": "Context Bytes",
            "type": "integer"
          },
          "context_tokens": {
            "maximum": 6000.0,
            "minimum": 0.0,
            "title": "Context Tokens",
            "type": "integer"
          },
          "corpus_id": {
            "pattern": "^corpus-sha256-[0-9a-f]{64}$",
            "title": "Corpus Id",
            "type": "string"
          },
          "index_id": {
            "pattern": "^index-sha256-[0-9a-f]{64}$",
            "title": "Index Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeHit"
            },
            "maxItems": 5,
            "title": "Items",
            "type": "array"
          },
          "provider": {
            "const": "fake",
            "default": "fake",
            "title": "Provider",
            "type": "string"
          },
          "retrieval_config_id": {
            "pattern": "^retrieval-config-sha256-[0-9a-f]{64}$",
            "title": "Retrieval Config Id",
            "type": "string"
          },
          "retrieval_version": {
            "const": "pgvector-cosine-exact-v1",
            "default": "pgvector-cosine-exact-v1",
            "title": "Retrieval Version",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "semantic_quality": {
            "const": "not_evaluated_fake_vectors",
            "default": "not_evaluated_fake_vectors",
            "title": "Semantic Quality",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "insufficient_evidence"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "index_id",
          "corpus_id",
          "retrieval_config_id",
          "items",
          "context_tokens",
          "context_bytes"
        ],
        "title": "RetrievalResult",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "RetailOps AI local access",
    "version": "0.1.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/access/forecast-check": {
      "post": {
        "operationId": "forecast_check_api_v1_access_forecast_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessDecision"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "408": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Timeout"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiBearer": []
          }
        ],
        "summary": "Forecast Check"
      }
    },
    "/api/v1/admin/access-policy": {
      "get": {
        "operationId": "policy_metadata_api_v1_admin_access_policy_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "408": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Timeout"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiBearer": []
          }
        ],
        "summary": "Policy Metadata"
      }
    },
...
    "/api/v1/identity": {
      "get": {
        "operationId": "identity_api_v1_identity_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "408": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Timeout"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiBearer": []
          }
        ],
        "summary": "Identity"
      }
    },
...
    "/api/v1/knowledge/search": {
      "post": {
        "operationId": "knowledge_search_api_v1_knowledge_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrievalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "408": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Timeout"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "apiBearer": []
          }
        ],
        "summary": "Knowledge Search"
      }
    }
  }
}